# openapi/apme.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apostille Me API",
    "version": "0.1.0",
    "description": "Public HTTP and WebSocket contract for apostille case intake, case lookup, case listing, health, readiness, metrics, and case-event streaming."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "operationId": "listCases",
        "summary": "List all in-memory apostille cases",
        "tags": [
          "cases"
        ],
        "description": "Case records can contain applicant and document metadata and are excluded from the baseline MCP catalog.",
        "responses": {
          "200": {
            "description": "Case list"
          }
        },
        "security": [],
        "x-ore-visibility": "public",
        "x-ore-stability": "stable",
        "x-ore-mcp-expose": false,
        "x-ore-mcp-mutating": false
      },
      "post": {
        "operationId": "createCase",
        "summary": "Create an apostille case",
        "tags": [
          "cases"
        ],
        "responses": {
          "201": {
            "description": "Case created"
          },
          "422": {
            "description": "Case input failed validation"
          }
        },
        "security": [],
        "x-ore-visibility": "public",
        "x-ore-stability": "stable",
        "x-ore-mcp-expose": false,
        "x-ore-mcp-mutating": true
      }
    },
    "/api/v1/cases/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Case UUID.",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getCase",
        "summary": "Get one apostille case",
        "tags": [
          "cases"
        ],
        "description": "Case records can contain applicant and document metadata and are excluded from the baseline MCP catalog.",
        "responses": {
          "200": {
            "description": "Case record"
          },
          "404": {
            "description": "Case not found"
          }
        },
        "security": [],
        "x-ore-visibility": "public",
        "x-ore-stability": "stable",
        "x-ore-mcp-expose": false,
        "x-ore-mcp-mutating": false
      }
    },
...
  }
}